api_comentarios/requirements.txt.txt
Flask==2.1.1
PyJWT==2.3.0
